Filter _oauth2_proxy cookie from litestar api logs
Modifies the built-in litestar StructlogPlugin to avoid printing the
cookie from oauth2-proxy. While oauth2-proxy supports customizing this
name in a way that would allow us to partially filter from output
(the cookie header would remain), the goal with this example project is
to demonstrate litestar functionality in a way that makes it easy to
jump in.

Note: also adds the __init__.py file to the api directory to prompt ruff
to treat this as a 1st party package for sorting imports correctly.